+++ packages/typespec-rust/test/spector/client/structure/two-operation-group/tests/two_operation_group_group2_client_test.rs
@@ -2,7 +2,7 @@
 //
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-use spector_twoop::{models::ClientType, TwoOperationGroupClient};
+use spector_twoop::{client::structure::service::models::ClientType, TwoOperationGroupClient};
 
 #[tokio::test]
 async fn five() {