+++ packages/typespec-rust/test/spector/parameters/body-optionality/src/lib.rs
@@ -7,3 +7,4 @@
 
 mod generated;
 pub use generated::*;
+pub mod optional_explicit;